Inhance performance for uploading image

+++ app/src/main/java/net/clipcodes/myapplication/utils/Libraries.java
@@ -103,11 +103,11 @@ public class Libraries {
         return bitmap;
     }
 
-    public static int uploadFile(String sourceFileUri, String fileRenamed) {
+    public static int uploadFile(String sourceFileUri, String fileRenamed, String copyPath) {
 
 
         int serverResponseCode = -1;
-        String fileName = sourceFileUri;
+//        String fileName = sourceFileUri;
         HttpURLConnection conn = null;
         DataOutputStream dos = null;
         String lineEnd = "\r\n";
@@ -118,7 +118,11 @@ public class Libraries {
         int maxBufferSize = 1 * 1024 * 1024;
 
         File sourceFile = new File(sourceFileUri);
-
+        Bitmap bitmap = Libraries.getBitmapFromFile(sourceFile);
+        Log.i("비트맵 사이즈(줄이기 전):", + bitmap.getHeight() * bitmap.getWidth() * 4 + " Byte");
+        sourceFile = Libraries.saveBitmapToJpeg(bitmap,  copyPath, Libraries.getFileName(sourceFileUri));
+        bitmap = Libraries.getBitmapFromFile(sourceFile);
+        Log.i("비트맵 사이즈(줄인 후):", bitmap.getHeight() * bitmap.getWidth() * 4 + " Byte");
         if (!sourceFile.isFile()) {
             return 0;
         }else{
@@ -144,7 +148,7 @@ public class Libraries {
 
                 conn.setRequestProperty("Content-Type", "multipart/form-data;boundary=" + boundary);
 
-                conn.setRequestProperty("uploaded_file", fileName);
+                conn.setRequestProperty("uploaded_file", sourceFileUri);
 
 
                 dos = new DataOutputStream(conn.getOutputStream());
@@ -265,6 +269,26 @@ public class Libraries {
 
     }
 
+    public static String getFileRootPath(String fileFullPath){
+        if(fileFullPath == null || fileFullPath.isEmpty()){
+            return null;
+        }
+
+        int startIdx = 0;
+        int endIdx = fileFullPath.indexOf(getFileName(fileFullPath));
+        fileFullPath = fileFullPath.substring(startIdx, endIdx);
+        return fileFullPath;
+    }
+
+    public static String getFileName(String fileFullPath){
+        if(fileFullPath == null || fileFullPath.isEmpty()){
+            return null;
+        }
+
+        File file = new File(fileFullPath);
+        return file.getName();
+    }
+
     public Bitmap readImageWithSampling(String imagePath, int targetWidth, int targetHeight) {
         // Get the dimensions of the bitmap
         BitmapFactory.Options bmOptions = new BitmapFactory.Options();
@@ -323,23 +347,32 @@ public class Libraries {
         }
         return null;
     }
-    public static void saveBitmapToJpeg(Bitmap bitmap, String cacheDir, String fileName) {
-
+    public static File saveBitmapToJpeg(Bitmap bitmap, String cacheDir, String fileName) {
+        File tempFile = null;
         boolean result = false;
 
+
         //storage 에 파일 인스턴스를 생성합니다.
-        File tempFile = new File(cacheDir, fileName);
+        tempFile = new File(cacheDir, fileName);
+        if(tempFile.exists()){
+           return tempFile;
+        }
 
         try {
 
             // 자동으로 빈 파일을 생성합니다.
+            tempFile.canWrite();
             result =  tempFile.createNewFile();
             Log.i("result: ", result+"");
             // 파일을 쓸 수 있는 스트림을 준비합니다.
             FileOutputStream out = new FileOutputStream(tempFile);
 
+            int width = bitmap.getWidth();
+            int height = bitmap.getHeight();
             // compress 함수를 사용해 스트림에 비트맵을 저장합니다.
-            result = bitmap.compress(Bitmap.CompressFormat.JPEG, 100, out);
+            bitmap = bitmap.createScaledBitmap(bitmap, 500, 500, true);
+            result = bitmap.compress(Bitmap.CompressFormat.JPEG, 30, out);
+
             Log.i("result: ", result+"");
             // 스트림 사용후 닫아줍니다.
             out.close();
@@ -349,6 +382,7 @@ public class Libraries {
         } catch (IOException e) {
             Log.e("MyTag","IOException : " + e.getMessage());
         }
+        return tempFile;
     }
 
     public static Bitmap getBitmapFromFile(File imgFile){
@@ -357,4 +391,5 @@ public class Libraries {
         }
         return null;
     }
+
 }
